Code Cleanup + remove old plugins support
